LIR/RealTimeDriverTracking/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
aiokafka
numpy
pydantic
pydantic-settings
python-dotenv
